+++ CHANGELOG.md
@@ -108,6 +108,7 @@
 - PIM-10152: Fix Table attribute option modal not scrolling on label translations
 - PIM-10120: Remove obsolete asset reference data
 - PIM-10151: Fix category field in rule edit form
+- PIM-10159: Fix table configuration update when a column is deleted and recreated with a different data type
 
 ## Improvements